requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
pymysql==1.1.1
pydantic==2.11.2
pydantic-settings==2.8.1
httpx==0.28.1
python-multipart==0.0.20
grpcio==1.71.0
grpcio-tools==1.71.0
pika==1.3.2
boto3==1.37.26
python-dateutil==2.9.0.post0
pytest==8.3.5

